src/main/webapp/app/entities/departamento-jefe/delete/departamento-jefe-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { DepartamentoJefeService } from '../service/departamento-jefe.service';

import { DepartamentoJefeDeleteDialogComponent } from './departamento-jefe-delete-dialog.component';

describe('DepartamentoJefe Management Delete Component', () => {
  let comp: DepartamentoJefeDeleteDialogComponent;
  let fixture: ComponentFixture<DepartamentoJefeDeleteDialogComponent>;
  let service: DepartamentoJefeService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [DepartamentoJefeDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(DepartamentoJefeDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(DepartamentoJefeDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(DepartamentoJefeService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
